Formatting, moving handicap dialog to his own file, adding handicap as a necessary tools for game_logic
